+++ spec/rubocop/config_spec.rb
@@ -224,6 +224,24 @@ describe RuboCop::Config do
           .to raise_error(RuboCop::ValidationError, message_matcher)
       end
     end
+
+    context 'when all cops are both Enabled and Disabled by default' do
+      before do
+        create_file(configuration_path, [
+                      'AllCops:',
+                      '  EnabledByDefault: true',
+                      '  DisabledByDefault: true'
+                    ])
+      end
+
+      it 'raises validation error' do
+        expect { configuration.validate }
+          .to raise_error(
+            RuboCop::ValidationError,
+            /Cops cannot be both enabled by default and disabled by default/
+          )
+      end
+    end
   end
 
   describe '#make_excludes_absolute' do
@@ -479,7 +497,11 @@ describe RuboCop::Config do
     end
   end
 
-  describe '#cop_enabled?' do
+  context 'whether the cop is enabled' do
+    def cop_enabled(cop_class)
+      configuration.for_cop(cop_class).fetch('Enabled')
+    end
+
     context 'when an entire cop department is disabled' do
       context 'but an individual cop is enabled' do
         let(:hash) do
@@ -491,7 +513,7 @@ describe RuboCop::Config do
 
         it 'still disables the cop' do
           cop_class = RuboCop::Cop::Style::TrailingWhitespace
-          expect(configuration.cop_enabled?(cop_class)).to be false
+          expect(cop_enabled(cop_class)).to be false
         end
       end
     end
@@ -507,7 +529,7 @@ describe RuboCop::Config do
 
         it 'still disables the cop' do
           cop_class = RuboCop::Cop::Style::TrailingWhitespace
-          expect(configuration.cop_enabled?(cop_class)).to be false
+          expect(cop_enabled(cop_class)).to be false
         end
       end
     end
@@ -521,7 +543,7 @@ describe RuboCop::Config do
 
       it 'enables the cop by default' do
         cop_class = RuboCop::Cop::Style::TrailingWhitespace
-        expect(configuration.cop_enabled?(cop_class)).to be true
+        expect(cop_enabled(cop_class)).to be true
       end
     end
   end